requirements.txt
langchain-community 
langchain-mcp-adapters
mcp
python-dotenv
gradio
openai
langgraph
langchain-openai
